Create Letter Point Values
store a list of correctly spelled words and a point value associated with each word
into a variable of type map<string, int>.

 Enter the point value from a the file called Letter point value.txt for each letter into a variable of type
vector<int>. The point value associated with `A' goes into position 0, and 'B' goes into position 1,
etc.
 The point value of a word is determined by the point value of each letter. To compute the point
value of a word, add up the point values of each letter in the word. For example, the point value of
"cat" is 4 + 1 + 1 = 6, since 'C' has a point value of 4, 'A' has a point value of 1, and 'T' has a
point value of 1. Upper and lower case letters have the same point value. (e.g. So "CAT" also has a
point value of 6.) Create a function to compute the point value of a word.